alternator: Store LSI keys in :attrs for newly created tables
Previously, LSI keys were stored as separate, top-level columns in the
base table. This patch changes this behavior for newly created tables,
so that the key columns are stored inside the `:attrs` map. Then, we use
top-level computed columns instead of regular ones.

This makes LSI storage consistent with GSIs and allows the use of a
collection tombstone on `:attrs` to delete all fields in a row, except
for keys in new tables.